Added update logic, added repo test for select

+++ src/main/scala/com.payments/domain/package.scala
@@ -2,6 +2,6 @@ package com.payments
 
 package object domain {
   final case class Name private (value: String)
-  final case class Amount private (value: Double)
+  final case class Amount private (value: BigDecimal)
   final case class Recipient private (value: String)
 }
